Appease everybody
